services/users/requirements.txt
blinker==1.4
Click==7.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
psycopg2-binary==2.7.6.1
SQLAlchemy==1.3.1
Werkzeug==0.14.1
coverage==4.5.2
flake8===3.6.0
flask-cors==3.0.7
Flask-Migrate==2.4.0
pyjwt==1.7.1